Supports and Resistances of a currency time signal
The goal of this work is to extract the supports (S) and resistances (R) statistically using a method that I will call it "Window size method". After extracting S and R their strengths will be evaluated using a statistical indicator. Support or resistance strength represents the possibility that the price is not able to exceed. If the S or R is strong, the probability that the price will reverse at this point is high. Otherwise we say S or R is weak.